Implement automatic update check (at 1-week interval).
