+++ .gitignore
@@ -15,3 +15,4 @@ __pycache__/
 .DS_Store
 /doc/
 .ipynb_checkpoints/
+/htmlcov/
