--- pyminflux/threads/_threads.py
@@ -1,30 +1,0 @@
-#  Copyright (c) 2022 - 2024 D-BSSE, ETH Zurich.
-#
-#  Licensed under the Apache License, Version 2.0 (the "License");
-#  you may not use this file except in compliance with the License.
-#  You may obtain a copy of the License at
-#
-#       http://www.apache.org/licenses/LICENSE-2.0
-#
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#  See the License for the specific language governing permissions and
-#   limitations under the License.
-#
-
-from PySide6 import QtCore
-
-
-class BaseThread(QtCore.QThread):
-    """
-    Thread that runs the tracker without blocking the UI.
-    """
-
-    def __init__(self, *args):
-        super().__init__()
-
-        self._args = args
-
-    def run(self):
-        print("Done.")
